Changed type to long

+++ src/Main.java
@@ -4,7 +4,7 @@
  */
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {19, 22, 312, 323, 25, 44};
+        long[] sales = {19, 22, 312, 323, 25, 44};
         SalesManager salesManager = new SalesManager(sales);
         System.out.println("Максимальное значение продаж: " + salesManager.max());
     }
